Fix accept and reject decorator issues

+++ labspy/labspaces/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.http import Http404, HttpResponseForbidden, HttpResponseNotAllowed
 from django.views.decorators.http import require_POST
+from django.urls import resolve
 
 from .models import Lab, LabMembership
 from .forms import LabCreationForm
@@ -45,7 +46,13 @@ def home(request):
 
 @login_required
 def lab_index(request):
-    joined_labs = LabMembership.objects.filter(user=request.user)
+    try:
+        joined_labs = LabMembership.objects.filter(user=request.user)
+    except LabMembership.DoesNotExist:
+        joined_labs = []
+    except Lab.DoesNotExist:
+        joined_labs = []
+    print("REQUEST:", request)
     return render(request, 'labspaces/lab_index.html', {'joined_labs': joined_labs})
 
 @login_required
@@ -70,6 +77,7 @@ def labspace_view(request, code):
         labspace = Lab.objects.get(code=code)  
     except Lab.DoesNotExist:
         raise Http404("Lab not found")
+    
     return render(request, 'labspaces/lab_index.html', {'labspace': labspace}) 
 
 @login_required
@@ -91,6 +99,8 @@ def lab_join(request):
 
 @login_required
 def pending_requests(request, code):
+    if not is_lab_admin(request.user, code):
+        return HttpResponseForbidden("You are not authorized to view this page")
     try:
         labspace = Lab.objects.get(code=code)
         pending_requests = LabMembership.objects.filter(lab=labspace, role='pending')
@@ -102,32 +112,36 @@ def pending_requests(request, code):
         raise Http404("Lab not found")
 
 @login_required
-@user_passes_test(is_lab_admin)
 @require_POST
 def accept_request(request, code):
-    try:
-        labspace = Lab.objects.get(code=code)
-        request_id = request.POST.get('request_id')
-        req = LabMembership.objects.get(id=request_id)
-        req.role = 'member'
-        req.save()
-        return redirect('labspaces:pending_requests', code=code)
-    except Lab.DoesNotExist:
-        raise Http404("Lab not found")
-    except LabMembership.DoesNotExist:
-        raise Http404("Request not found")
+    if is_lab_admin(request.user, code):
+        try:
+            labspace = Lab.objects.get(code=code)
+            request_id = request.POST.get('request_id')
+            req = LabMembership.objects.get(id=request_id)
+            req.role = 'member'
+            req.save()
+            return redirect('labspaces:pending_requests', code=code)
+        except Lab.DoesNotExist:
+            raise Http404("Lab not found")
+        except LabMembership.DoesNotExist:
+            raise Http404("Request not found")
+    else: 
+        return HttpResponseForbidden("You do not have permission to perform this action.")
 
 @login_required
-@user_passes_test(is_lab_admin)
 @require_POST
 def reject_request(request, code):
-    try:
-        labspace = Lab.objects.get(code=code)
-        user_id = request.POST.get('pending_user')
-        pending_user = LabMembership.objects.get(id=user_id)
-        pending_user.delete()
-        return redirect('labspaces:pending_requests', code=code)
-    except Lab.DoesNotExist:
-        raise Http404("Lab not found")
-    except LabMembership.DoesNotExist:
-        raise Http404("Request not found")
+    if is_lab_admin(request.user, code):
+        try:
+            labspace = Lab.objects.get(code=code)
+            user_id = request.POST.get('pending_user')
+            pending_user = LabMembership.objects.get(id=user_id)
+            pending_user.delete()
+            return redirect('labspaces:pending_requests', code=code)
+        except Lab.DoesNotExist:
+            raise Http404("Lab not found")
+        except LabMembership.DoesNotExist:
+            raise Http404("Request not found")
+    else:
+        return HttpResponseForbidden("You do not have permission to perform this action.")
